Sprint 08 Review changes
- Add table view constraints
- Reorganize extensions
- Update access modifiers
- Update toolbar style
